+++ app/src/main/java/seamuslowry/daytracker/data/repos/SettingsRepo.kt
@@ -25,8 +25,6 @@ class SettingsRepo @Inject constructor(@ApplicationContext private val context: 
         val REMINDER_ENABLED = booleanPreferencesKey("REMINDER_ENABLED_KEY")
         val REMINDER_TIME = stringPreferencesKey("REMINDER_TIME")
         val SHOW_RECORDED_VALUES = booleanPreferencesKey("SHOW_RECORDED_VALUES")
-        val LOW_VALUE_COLOR = intPreferencesKey("LOW_VALUE_COLOR")
-        val HIGH_VALUE_COLOR = intPreferencesKey("HIGH_VALUE_COLOR")
     }
 
     suspend fun setReminderEnabled(enabled: Boolean) {
@@ -47,38 +45,12 @@ class SettingsRepo @Inject constructor(@ApplicationContext private val context: 
         }
     }
 
-    suspend fun setLowValueColor(color: Color) {
-        context.dataStore.edit {
-            it[LOW_VALUE_COLOR] = color.toArgb()
-        }
-    }
-
-    suspend fun setHighValueColor(color: Color) {
-        context.dataStore.edit {
-            it[HIGH_VALUE_COLOR] = color.toArgb()
-        }
-    }
-
     val settings: Flow<Settings> = context.dataStore.data
         .map {
             Settings(
                 reminderEnabled = it[REMINDER_ENABLED] == true,
                 reminderTime = it[REMINDER_TIME]?.let { time -> LocalTime.parse(time) } ?: LocalTime.of(18, 0),
                 showRecordedValues = it[SHOW_RECORDED_VALUES] == true,
-                lowValueColor = it[LOW_VALUE_COLOR]?.let { colorInt ->
-                    try {
-                        Color(colorInt)
-                    } catch (_: Exception) {
-                        null
-                    }
-                },
-                highValueColor = it[HIGH_VALUE_COLOR]?.let { colorInt ->
-                    try {
-                        Color(colorInt)
-                    } catch (_: Exception) {
-                        null
-                    }
-                },
             )
         }
 }
@@ -87,6 +59,4 @@ data class Settings(
     val reminderEnabled: Boolean = false,
     val reminderTime: LocalTime = LocalTime.of(18, 0),
     val showRecordedValues: Boolean = false,
-    val lowValueColor: Color? = null,
-    val highValueColor: Color? = null,
 )
