+++ app/src/main/java/seamuslowry/daytracker/ui/screens/settings/SettingsScreen.kt
@@ -4,24 +4,14 @@ import android.Manifest
 import android.content.pm.PackageManager
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Restore
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
@@ -37,7 +27,6 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.contentDescription
@@ -48,9 +37,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import kotlinx.coroutines.launch
 import seamuslowry.daytracker.R
 import seamuslowry.daytracker.models.localeFormat
-import seamuslowry.daytracker.models.toHexString
 import java.time.LocalTime
-import androidx.core.graphics.toColorInt
 
 @Composable
 fun SettingsScreen(
@@ -59,9 +46,6 @@ fun SettingsScreen(
     val state by viewModel.state.collectAsState()
 
     val scope = rememberCoroutineScope()
-
-    val defaultHighColor = MaterialTheme.colorScheme.primary
-    val defaultLowColor = MaterialTheme.colorScheme.error
 
     Column(
         modifier = Modifier
@@ -76,19 +60,6 @@ fun SettingsScreen(
             onSetReminderTime = { scope.launch { viewModel.setReminderTime(it) } },
         )
         CalendarSection(showValues = state.showRecordedValues, onSetShowValues = { scope.launch { viewModel.setShowRecordedValues(it) } })
-        ColorSection(
-            lowColor = state.lowValueColor ?: defaultLowColor,
-            highColor = state.highValueColor ?: defaultHighColor,
-            onSetLowColor = { scope.launch { viewModel.setLowValueColor(it) } },
-            onSetHighColor = { scope.launch { viewModel.setHighValueColor(it) } },
-            onReset = {
-                scope.launch {
-                    viewModel.setLowValueColor(defaultLowColor)
-                    viewModel.setHighValueColor(defaultHighColor)
-                }
-            },
-            allowReset = state.lowValueColor != defaultLowColor || state.highValueColor != defaultHighColor,
-        )
     }
 }
 
@@ -172,98 +143,6 @@ fun ReminderSection(
             onDismiss = { pickingTime = false },
         )
     }
-}
-
-@Composable
-fun ColorSection(
-    lowColor: Color,
-    highColor: Color,
-    onSetLowColor: (c: Color) -> Unit,
-    onSetHighColor: (c: Color) -> Unit,
-    onReset: () -> Unit,
-    allowReset: Boolean,
-    modifier: Modifier = Modifier,
-) {
-    Column(modifier = modifier) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Text(
-                text = stringResource(R.string.color_section_title),
-                modifier = Modifier.padding(vertical = 8.dp),
-                style = MaterialTheme.typography.headlineSmall,
-            )
-            if (allowReset) {
-                IconButton(onClick = onReset) {
-                    Icon(imageVector = Icons.Filled.Restore, contentDescription = stringResource(R.string.reset_to_defaults))
-                }
-            }
-        }
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 4.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Text(text = stringResource(R.string.low_color), modifier = Modifier.weight(1f))
-            ColorTextField(color = lowColor, onColorChange = onSetLowColor, modifier = Modifier.weight(1f))
-        }
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(top = 4.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween,
-        ) {
-            Text(text = stringResource(R.string.high_color), modifier = Modifier.weight(1f))
-            ColorTextField(color = highColor, onColorChange = onSetHighColor, modifier = Modifier.weight(1f))
-        }
-    }
-}
-
-@Composable
-private fun ColorTextField(
-    color: Color,
-    onColorChange: (c: Color) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    var textColor by remember { mutableStateOf(color.toHexString()) }
-
-    LaunchedEffect(key1 = textColor) {
-        try {
-            onColorChange(Color("#$textColor".toColorInt()))
-        } catch (_: Exception) { }
-    }
-
-    LaunchedEffect(key1 = color) {
-        textColor = color.toHexString()
-    }
-
-    OutlinedTextField(
-        value = textColor,
-        onValueChange = { textColor = it.uppercase().take(6) },
-        modifier = modifier,
-        prefix = { Text(text = stringResource(R.string.hex_prefix)) },
-        trailingIcon = {
-            Box(
-                modifier = Modifier
-                    .size(24.dp)
-                    .background(
-                        color = if (color.toHexString() == textColor) color else Color.Unspecified,
-                        shape = CircleShape,
-                    )
-                    .border(
-                        width = 1.dp,
-                        color = MaterialTheme.colorScheme.outline,
-                        shape = CircleShape,
-                    ),
-            )
-        },
-    )
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
